start implementing memory
