+++ src/lib.rs
@@ -2,3 +2,4 @@ pub mod chains;
 pub mod llm;
 pub mod prompt;
 pub mod record;
+pub mod memory;
